% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive.r
\name{add_derived}
\alias{add_derived}
\title{Add a derived variable to the dictionary}
\usage{
add_derived(dict, var, fun)
}
\arguments{
\item{dict}{A variable dictionary returned by
\code{\link[qualtdict]{dict_generate}}.}

\item{var}{String. Name of the derived variable to be added.}

\item{fun}{Function. A function that takes a survey data frame as input
and returns the value of the derived variable.}
}
\description{
Add a derived variable to the dictionary. An additional \code{formula}
column is created to store a function that deriives this variable.
}
