% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{checkarg_isfunction}
\alias{checkarg_isfunction}
\title{Is one of \code{c("question_name", "easy_name")}}
\usage{
checkarg_isfunction(arg, null_okay = TRUE)
}
\description{
Is one of \code{c("question_name", "easy_name")}
}
\keyword{internal}
