% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{checkarg_isname}
\alias{checkarg_isname}
\title{Is one of \code{c("question_name", "easy_name")}}
\usage{
checkarg_isname(arg)
}
\description{
Is one of \code{c("question_name", "easy_name")}
}
\keyword{internal}
