% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_html}
\alias{convert_html}
\title{Convert html special characeters}
\usage{
convert_html(data)
}
\arguments{
\item{data}{A dataframe}
}
\description{
Convert html special characeters
}
\keyword{internal}
