% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dict_upload.r
\name{data_update}
\alias{data_update}
\title{Update a survey data frame by adding derived variables from the
dictionary}
\usage{
data_update(data, dict)
}
\arguments{
\item{data}{Data frame. A survey data frame returned by
\code{\link[qualtdict]{get_survey_data}}}

\item{dict}{A variable dictionary returned by
\code{\link[qualtdict]{dict_generate}}.}
}
\description{
Add values of derived variables from a dictionary to the survey data
frame.
}
