% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dict_compare.r
\name{dict_compare}
\alias{dict_compare}
\title{Compare two dictionaries and suggest potential matching variables}
\usage{
dict_compare(dict, reference_dict, field = c("all", "question", "item"), ...)
}
\arguments{
\item{dict}{A variable dictionary returned by
\code{\link[qualtdict]{dict_generate}}.}

\item{reference_dict}{Variable dictionary returned by
\code{\link[qualtdict]{dict_generate}}. The variable names in this
dictionary wil be used when the object returned by the function is used
for \code{\link[qualtdict]{dict_merge}} or \code{\link[qualtdict]{dict_rename}}}

\item{field}{String. Which field is used when comparing variables.}

\item{...}{Other arguments passed to
\code{\link[stringdist]{amatch}} to configure fuzzy matching.}
}
\description{
Compare variables in two dictionaries based on either their question or
item text and labels and suggest potential (fuzzy) matches by comparing
each question or item text in \code{dict} to all the ones in
\code{reference_dict} and obtain the best match.
The results can be
used for
\code{\link[qualtdict]{dict_rename}} to ensure the same variables in
two different dictionaries have the same names.
}
