% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dict_upload.r
\name{dict_download}
\alias{dict_download}
\title{Download dictionary from a googlesheet with \code{surveyID} and
\code{survey_name} metadata}
\usage{
dict_download(url)
}
\arguments{
\item{url}{String. An url for a googlesheet page.}
}
\description{
Download dictionary to a googlesheet. \code{surveyID} and
\code{survey_name} are added to the dictionary as attirbutes.
}
