% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dict_upload.r
\name{dict_upload}
\alias{dict_upload}
\title{Upload dictionary to a googlesheet with \code{surveyID} and
\code{survey_name} metadata kept}
\usage{
dict_upload(dict, url, surveyID = FALSE)
}
\arguments{
\item{dict}{A variable dictionary returned by
\code{\link[qualtdict]{dict_generate}}.}

\item{url}{String. An url for a googlesheet page.}

\item{surveyID}{Logical. Whether to write \code{surveyID} to the
googlesheet.}
}
\description{
Upload dictionary to a googlesheet. \code{surveyID} and
\code{survey_name} are added to the sheet.
}
