% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyname_gen.R
\name{easyname_gen}
\alias{easyname_gen}
\title{Generate easy names from dictionary}
\usage{
easyname_gen(json, surveyID, block_pattern, block_sep, preprocess)
}
\description{
Generate easy names from dictionary
}
\keyword{internal}
