% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fetch_description2}
\alias{fetch_description2}
\title{Retry version of \code{fetch_description}}
\usage{
fetch_description2(...)
}
\description{
Retry version of \code{fetch_description}
}
\keyword{internal}
