% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_onetoone}
\alias{is_onetoone}
\title{Given a two-column dataframe determine which row is one-to-one (logical)}
\usage{
is_onetoone(cols)
}
\arguments{
\item{cols}{A dataframe with two columns}
}
\description{
Given a two-column dataframe determine which row is one-to-one (logical)
}
\keyword{internal}
