% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labelled_export.R
\name{labelled_export_findings}
\alias{labelled_export_findings}
\title{Return Labelled Export Findings}
\usage{
labelled_export_findings(x)
}
\arguments{
\item{x}{Labelled Survey Data returned by \code{\link[=get_survey_data]{get_survey_data()}}.}
}
\value{
A data frame of Labelled Export Findings.
}
\description{
Labelled Export Findings describe issues detected while matching a Variable
Dictionary to downloaded survey data.
}
