% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{null_na}
\alias{null_na}
\title{Convert NULL or a list of NULLs to NA}
\usage{
null_na(x)
}
\arguments{
\item{x}{NULL or a list of NULLs}
}
\description{
Convert NULL or a list of NULLs to NA
}
\keyword{internal}
