% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualtdict-package.R
\docType{package}
\name{qualtdict-package}
\alias{qualtdict}
\alias{qualtdict-package}
\title{qualtdict: Generate Variable Dictionaries and Labelled Survey Data for 'Qualtrics' Surveys}
\description{
Generates Variable Dictionaries from 'Qualtrics' \url{https://www.qualtrics.com/about/} survey metadata and applies those dictionaries to create Labelled Survey Data. This package is a metadata and labelled-export companion to 'qualtRics', which provides access to 'Qualtrics' survey data and metadata via the 'Qualtrics' API \url{https://api.qualtrics.com/}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/lyh970817/qualtdict}
  \item \url{https://lyh970817.github.io/qualtdict/}
  \item Report bugs at \url{https://github.com/lyh970817/qualtdict/issues}
}

}
\author{
\strong{Maintainer}: Yuhao Lin \email{yuhao.lin@kcl.ac.uk} (\href{https://orcid.org/0000-0001-6357-5731}{ORCID})

Authors:
\itemize{
  \item Yuhao Lin \email{yuhao.lin@kcl.ac.uk} (\href{https://orcid.org/0000-0001-6357-5731}{ORCID})
}

}
\keyword{internal}
