% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_json.R
\name{recode_json}
\alias{recode_json}
\title{Download metadata in json format and recode into dictionary format}
\usage{
recode_json(surveyID, easyname_gen, block_pattern, block_sep, preprocess)
}
\description{
Download metadata in json format and recode into dictionary format
}
\keyword{internal}
