% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{retry}
\alias{retry}
\title{Function factory for maximum five times of retries when failed}
\usage{
retry(f)
}
\arguments{
\item{f}{A function}
}
\description{
Function factory for maximum five times of retries when failed
}
\keyword{internal}
