% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_survey_data.R
\name{survey_recode}
\alias{survey_recode}
\title{Add labels to survey}
\usage{
survey_recode(dict, dat, keys, unanswer_recode, unanswer_recode_multi)
}
\description{
Add labels to survey
}
\keyword{internal}
