% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_survey_data.R
\name{survey_var_recode}
\alias{survey_var_recode}
\title{Add labels to each variable in survey (\code{sjlabelled} uses \code{haven})}
\usage{
survey_var_recode(var, var_dict, unanswer_recode, unanswer_recode_multi)
}
\description{
Add labels to each variable in survey (\code{sjlabelled} uses \code{haven})
}
\keyword{internal}
