% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{unique_expand}
\alias{unique_expand}
\title{Suppose x = f(unique(y)), find f(y)}
\usage{
unique_expand(x, y)
}
\arguments{
\item{x}{A character vector}

\item{y}{A character vector}
}
\description{
Suppose x = f(unique(y)), find f(y)
}
\keyword{internal}
