% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{which_not_onetoone}
\alias{which_not_onetoone}
\title{Given a two-column dataframe find which row is not one-to-one}
\usage{
which_not_onetoone(cols)
}
\arguments{
\item{cols}{A dataframe with two columns}
}
\description{
Given a two-column dataframe find which row is not one-to-one
}
\keyword{internal}
