% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dict_generate.R
\name{dict_generate}
\alias{dict_generate}
\title{Generate a variable dictionary for a survey}
\usage{
dict_generate(
  surveyID,
  name = c("question_name", "easy_name"),
  block_pattern = NULL,
  block_sep = ".",
  preprocess = NULL,
  split_by_block = FALSE,
  survey_name = NULL
)
}
\arguments{
\item{surveyID}{String. Unique ID for the survey you want to download.}

\item{name}{String. Which type of variable names to use in the dictionary,
\code{question_name} for question names on Qualtrics, \code{easy_name}
for easy names generated based on question text.}

\item{block_pattern}{Function. A function that given the name of a
survey block, returns a string to be appended to variable names
in that block.
Defaults to \code{NULL}.}

\item{block_sep}{String. Seperator between variable names and block
prefixes returned by \code{block_pattern}. Defaults to ".".}

\item{preprocess}{Function. A function that given a dataframe with the
column names as in a dictionary, does formatting for each column.
Defaults to \code{NULL}.}

\item{split_by_block}{Logical. If \code{TRUE}, the function returns a
list with each element being the dictionary for a single survey block.}

\item{survey_name}{String Specified to add to attribute
\code{survey_name}.}
}
\value{
A dictionary dataframe or a list of dictionary dataframes (depending on
\code{split_by_block}.
}
\description{
Generate a variable dictionary for a Qualtrics survey survey
}
\examples{
\dontrun{

# Create a function for \code{block_pattern}
# that returns the first three letters of a string
block_pattern <- function(x) {
  substring(x, 1, 3)
}


mydict <- dict_generate("SV_4YyAHbAxpdbzacl",
  name = "easy_name",
  block_pattern = block_pattern,
  block_sep = ".",
  split_by_block = FALSE
)
}

}
