% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dict_validate.r
\name{dict_validate}
\alias{dict_validate}
\title{Check for potential mistakes in the dictionary}
\usage{
dict_validate(dict)
}
\arguments{
\item{dict}{A variable dictionary returned by
\code{\link[qualtdict]{dict_generate}}.}
}
\description{
Check for potential mistakes in the dictionary: 1) Whether the variable
names are unique; 2) For each variable, whether there is a unique
mapping between level and label and whether the levels are consecutive;
and extract unique level-label pairings.
}
