% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_survey_data.r
\name{get_survey_data}
\alias{get_survey_data}
\title{Download a labeled survey data set}
\usage{
get_survey_data(
  dict,
  surveyID = NULL,
  keys = NULL,
  split_by_block = FALSE,
  skip_mistakes = FALSE,
  ...
)
}
\arguments{
\item{dict}{A variable dictionary returned by
\code{\link[qualtdict]{dict_generate}}.}

\item{surveyID}{String. A variable dictionary returned by
\code{\link[qualtdict]{dict_generate}} has the survey ID as an
attribute. If it is read from elsewhere, this needs to be specified
manually. Defaults to \code{NULL}.}

\item{keys}{A character vector containing variables to be added, if
\code{split_by_block} is \code{TRUE}, to all individual block data sets.
Can also be used to add variables (e.g. IP address) found on Qualtrics
but not in the dictionary to the downloaded data sets.}

\item{split_by_block}{Logical. If \code{TRUE}, the function returns a
list with each element being the data set for a single survey block.}

\item{skip_mistakes}{Logical. If \code{TRUE}, variables with potenetial
level-label mistakes will be removed from the data set.}

\item{...}{Other arguments passed to
\code{\link[qualtRics]{fetch_survey}}. Note that \code{surveyID},
\code{import_id}, \code{convert}, \code{label} and \code{include_qids}
will be overwritten by the function.}
}
\description{
Download a survey data set from Qualtrics corresponding to a variable
dictionary generated by \code{\link[qualtdict]{dict_generate}}.
Question, item texts and level labels are added as attributes.
}
\examples{
\dontrun{

# Generate a dictionary
mydict <- dict_generate("SV_4YyAHbAxpdbzacl",
  survey_name = "mysurvey",
  var_name = "easy_name",
  block_pattern = block_pattern,
  block_sep = ".",
  split_by_block = FALSE
)
survey_dat <- get_survey_data(mydict,
  unanswer_recode = -77,
  unanswer_recode_multi = 0
)
}
}
